Refactor into visitor and formatter
This refactors the visitor and URI template building to be faster, simpler
and easier to test. It should be easier to add explicit caching too.

The new RFC6570::Visitor only traverses the Journey AST to produces a
list of parts needed to build an URI template. It does not filter for ignored
substitutions nor append controller parameters to the part list. This
means we can run this once for a route and reuse the much simpler and faster
parts list for build customizable URI templates. The visitor uses its cached
method dispatch for dual-node CAT clauses too, simplifying the implementation.
Last but not least it got its own test suite too.

The new RFC6570::Formatter uses the part list from a visitor to build URI
templates. It filters for ignored substitutions, adds action parameters and
contains all code needed to build path only or full URI templates. It further
avoids creating templates twice when generating full URLs.